+++ lib/boruta/oauth/request/base.ex
@@ -130,11 +130,13 @@ defmodule Boruta.Oauth.Request.Base do
       client_id: params["client_id"],
       resource_owner: params["resource_owner"],
       redirect_uri: params["redirect_uri"],
+      relying_party_redirect_uri: params["relying_party_redirect_uri"],
       state: params["state"],
       nonce: params["nonce"],
       prompt: params["prompt"],
       code_challenge: params["code_challenge"],
       code_challenge_method: params["code_challenge_method"],
+      code: params["code"],
       scope: params["scope"],
       client_metadata: client_metadata,
       response_type: params["response_type"]